Added helper functions to Limelight subsystem
Instead of returning a LimelightTransform (which was removed), the functions to get the robot position and the relative position of a tag return more usable Pose2d objects
Also added the AprilTag enum class to more easily reference different AprilTags in code. Their pose variable is taken from the data provided on the Limelight docs for field orientation. Rotations *should* be the rotation a robot needs to be in to have the front facing the marker. Needs testing